chore: merge server in one
