+++ apps/blog/src/app/(pages)/[post]/opengraph-image.tsx
@@ -1,6 +1,6 @@
 import { ImageResponse } from "next/og";
 import { logger, monitor } from "@ashgw/observability";
-import { httpClient } from "~/trpc/callers/server/http";
+import { trpcHttpServerSideClient } from "~/trpc/callers/server/http";
 
 export const runtime = "nodejs";
 
@@ -25,9 +25,10 @@ const clamp = (s: string, n: number) =>
 
 export default async function Image(_: Request, { params }: RouteCtx) {
   try {
-    const post = await httpClient.post.getDetailedPublicPost.query({
-      slug: params.post,
-    });
+    const post =
+      await trpcHttpServerSideClient.post.getDetailedPublicPost.query({
+        slug: params.post,
+      });
 
     const title = clamp(post?.title ?? "Post not found", 90);
     const subtitle = clamp(post?.summary ?? "No description available", 140);
