+++ apps/blog/src/app/components/pages/home/index.tsx
@@ -1,9 +1,9 @@
 import { HydrateClient } from "~/trpc/server";
 import { PostCardsPage } from "./components/PostCardsPage";
-import { httpClient } from "~/trpc/callers/server/http";
+import { trpcHttpServerSideClient } from "~/trpc/callers/server/http";
 
 export async function HomePage() {
-  const posts = await httpClient.post.getPublicPostCards.query();
+  const posts = await trpcHttpServerSideClient.post.getPublicPostCards.query();
   return (
     <HydrateClient>
       <PostCardsPage posts={posts} />
